requirements.txt
python-dotenv
gpio
pi1wire